% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_path.R
\name{get_mrc_path}
\alias{get_mrc_path}
\title{Get the full path of a \code{mauricer} file}
\usage{
get_mrc_path(filename)
}
\arguments{
\item{filename}{the file's name, without the path}
}
\value{
the full path of the filename, if and only if
  the file is present. Will stop otherwise.
}
\description{
Get the full path of a file in the \code{inst/extdata} folder.
If there is no \code{mauricer} file, \link{get_mrc_path} will \link{stop}.
}
\examples{
get_mrc_path("anthus_aco_sub.fas")
}
\seealso{
for more files, use \code{\link{get_mrc_paths}}
}
\author{
Richèl J.C. Bilderbeek
}
