% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_paths.R
\name{get_mrc_paths}
\alias{get_mrc_paths}
\title{Get the full path of one or more \code{mauricer} files}
\usage{
get_mrc_paths(filenames)
}
\arguments{
\item{filenames}{the files' names, without the path}
}
\value{
the filenames' full paths
}
\description{
Get the full paths of files in the \code{inst/extdata} folder
If there is a \code{mauricer} file absent,
\link{get_mrc_paths} will \link{stop}.
}
\examples{
get_mrc_paths(c("anthus_aco_sub.fas", "anthus_nd2_sub.fas"))
}
\seealso{
for one file, use \code{\link{get_mrc_path}}
}
\author{
Richèl J.C. Bilderbeek
}
