% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_internet.R
\name{has_internet}
\alias{has_internet}
\title{Check to see if there is an internet connection}
\usage{
has_internet()
}
\value{
TRUE if there is an internet connection, FALSE otherwise
}
\description{
Used code from https://stackoverflow.com/a/34970085
}
\author{
Richèl J.C. Bilderbeek copied the code here from
  the original author Jeroen Ooms
}
