% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_beast2_package.R
\name{install_beast2_package}
\alias{install_beast2_package}
\title{Install a BEAST2 package}
\usage{
install_beast2_package(name)
}
\arguments{
\item{name}{the package's name}
}
\value{
nothing. It does install the BEAST2 package
}
\description{
Install a BEAST2 package
}
