% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_beast2_ns_pkg_installed.R
\name{is_beast2_ns_pkg_installed}
\alias{is_beast2_ns_pkg_installed}
\title{Is the BEAST2 NS package installed?}
\usage{
is_beast2_ns_pkg_installed(
  show_warnings = FALSE,
  verbose = FALSE,
  beast2_path = beastier::get_default_beast2_path()
)
}
\arguments{
\item{show_warnings}{if TRUE, warnings will shown}

\item{verbose}{if TRUE, additional information is displayed, that
is potentially useful in debugging}

\item{beast2_path}{name of either a BEAST2 binary file
(usually simply \code{beast})
or a BEAST2 jar file
(usually has a \code{.jar} extension).
Use \link[beastier]{get_default_beast2_bin_path} to get
the default BEAST binary file's path
Use \link[beastier]{get_default_beast2_jar_path} to get
the default BEAST jar file's path}
}
\value{
TRUE if the BEAST2 NS package is installed, FALSE otherwise
}
\description{
Determine if the BEAST2 NS package is installed.
}
\details{
Unlike \link{is_beast2_pkg_installed},
this function does not need an internet connection.
Instead, the function calls BEAST2 to read a BEAST2 XML file that
uses NS.
}
\examples{
\dontrun{
  is_beast2_ns_pkg_installed()
}
}
\author{
Richèl J.C. Bilderbeek
}
