% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_beast2_packages.R
\name{list_beast2_packages}
\alias{list_beast2_packages}
\title{List all BEAST2 packages available and installed}
\usage{
list_beast2_packages()
}
\value{
a data frame
}
\description{
List all BEAST2 packages available and installed
}
\examples{
  df <- list_beast2_packages()
  testthat::expect_true("name" \%in\% names(df))
  testthat::expect_true("installed_version" \%in\% names(df))
  testthat::expect_true("latest_version" \%in\% names(df))
  testthat::expect_true("dependencies" \%in\% names(df))
  testthat::expect_true("description" \%in\% names(df))
}
\author{
Richel J.C. Bilderbeek
}
