% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrc_install.R
\name{mrc_install}
\alias{mrc_install}
\title{Install a BEAST2 package}
\usage{
mrc_install(name)
}
\arguments{
\item{name}{the package's name}
}
\value{
nothing. It does install the BEAST2 package
}
\description{
Install a BEAST2 package
}
