% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrc_is_installed.R
\name{mrc_is_installed}
\alias{mrc_is_installed}
\title{Checks if a BEAST2 package is already installed}
\usage{
mrc_is_installed(name)
}
\arguments{
\item{name}{BEAST2 package name}
}
\value{
TRUE if the BEAST2 package is installed
}
\description{
Checks if a BEAST2 package is already installed
}
\author{
Richel J.C. Bilderbeek
}
