% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_beast2_pkg_names.R
\name{mrc_list}
\alias{mrc_list}
\title{List all BEAST2 packages available and installed}
\usage{
mrc_list()
}
\value{
a data frame
}
\description{
List all BEAST2 packages available and installed
}
\author{
Richel J.C. Bilderbeek
}
