% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrc_uninstall.R
\name{mrc_uninstall}
\alias{mrc_uninstall}
\title{Uninstall a BEAST2 package}
\usage{
mrc_uninstall(name)
}
\arguments{
\item{name}{the package's name}
}
\value{
nothing. It does install the BEAST2 package
}
\description{
Uninstall a BEAST2 package
}
\author{
Richel J.C. Bilderbeek
}
