% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_beast2_pkg.R
\name{install_beast2_pkg}
\alias{install_beast2_pkg}
\title{Install a BEAST2 package}
\usage{
install_beast2_pkg(name)
}
\arguments{
\item{name}{the package's name}
}
\value{
nothing. It does install the BEAST2 package
}
\description{
Install a BEAST2 package
}
\examples{
  if (beastier::is_beast2_installed() && !is_beast2_pkg_installed("NS")) {
    install_beast2_pkg("NS")
  }
}
\author{
Richèl J.C. Bilderbeek
}
