% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_beast2_pkg.R
\name{install_beast2_pkg}
\alias{install_beast2_pkg}
\title{Install a BEAST2 package}
\usage{
install_beast2_pkg(name)
}
\arguments{
\item{name}{the package's name}
}
\value{
nothing. It does install the BEAST2 package
}
\description{
Install a BEAST2 package. If the package is already installed,
(see \link{is_beast2_pkg_installed}), this function \link{stop}s.
}
\examples{
\dontrun{
  install_beast2_pkg("NS")
}
}
\author{
Richèl J.C. Bilderbeek
}
