% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_beast2_ns_pkg_installed.R
\name{is_beast2_ns_pkg_installed}
\alias{is_beast2_ns_pkg_installed}
\title{Determine if the BEAST2 NS package is installed}
\usage{
is_beast2_ns_pkg_installed()
}
\value{
TRUE if the BEAST2 NS package is installed, FALSE otherwise
}
\description{
Determine if the BEAST2 NS package is installed
}
\examples{
library(testthat)

if (is_beast2_installed()) {
  print(
    paste(
      "Is the BEAST2 NS package installed:",
      is_beast2_ns_pkg_installed()
    )
  )
}
}
\author{
Richèl J.C. Bilderbeek
}
