% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_beast2_ns_pkg_installed.R
\name{is_beast2_ns_pkg_installed}
\alias{is_beast2_ns_pkg_installed}
\title{Is the BEAST2 NS package installed?}
\usage{
is_beast2_ns_pkg_installed()
}
\value{
TRUE if the BEAST2 NS package is installed, FALSE otherwise
}
\description{
Determine if the BEAST2 NS package is installed.
}
\details{
Unlike \link{is_beast2_pkg_installed},
this function does not need an internet connection.
Instead, the function calls BEAST2 to read a BEAST2 XML file that
uses NS.
}
\examples{
\dontrun{
  is_beast2_ns_pkg_installed()
}
}
\author{
Richèl J.C. Bilderbeek
}
