% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_beast2_pkg_installed.R
\name{is_beast2_pkg_installed}
\alias{is_beast2_pkg_installed}
\title{Checks if a BEAST2 package is already installed}
\usage{
is_beast2_pkg_installed(name)
}
\arguments{
\item{name}{BEAST2 package name}
}
\value{
TRUE if the BEAST2 package is installed
}
\description{
Checks if a BEAST2 package is already installed
}
\examples{
  library(testthat)

  if (is_beast2_installed()) {
    print(
      paste(
        "Is the NS BEAST2 package installed:",
        is_beast2_pkg_installed("NS")
      )
    )
  }
}
\author{
Richèl J.C. Bilderbeek
}
