% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_beast2_packages.R
\name{list_beast2_packages}
\alias{list_beast2_packages}
\title{List all BEAST2 packages available and installed}
\usage{
list_beast2_packages()
}
\value{
a data frame
}
\description{
List all BEAST2 packages available and installed
}
\author{
Richel J.C. Bilderbeek
}
