% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_beast2_pkg.R
\name{mrc_install}
\alias{mrc_install}
\title{Obsolete function to install a BEAST2 package.
Use \link{install_beast2_pkg} instead}
\usage{
mrc_install(name = NA)
}
\arguments{
\item{name}{the package's name}
}
\value{
nothing. It does install the BEAST2 package
}
\description{
Obsolete function to install a BEAST2 package.
Use \link{install_beast2_pkg} instead
}
\author{
Richèl J.C. Bilderbeek
}
