% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_beast2_pkg_installed.R
\name{mrc_is_installed}
\alias{mrc_is_installed}
\title{Obsolete function to checks if a BEAST2 package is already installed.
Use \link{is_beast2_pkg_installed} instead}
\usage{
mrc_is_installed(name = NA)
}
\arguments{
\item{name}{BEAST2 package name}
}
\value{
TRUE if the BEAST2 package is installed
}
\description{
Obsolete function to checks if a BEAST2 package is already installed.
Use \link{is_beast2_pkg_installed} instead
}
\author{
Richèl J.C. Bilderbeek
}
