% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrc_list.R
\name{mrc_list}
\alias{mrc_list}
\title{List all BEAST2 packages available and installed}
\usage{
mrc_list()
}
\value{
a data frame
}
\description{
List all BEAST2 packages available and installed
}
\examples{
  df <- mrc_list()
  testthat::expect_true("name" \%in\% names(df))
  testthat::expect_true("installed_version" \%in\% names(df))
  testthat::expect_true("latest_version" \%in\% names(df))
  testthat::expect_true("dependencies" \%in\% names(df))
  testthat::expect_true("description" \%in\% names(df))
}
\author{
Richel J.C. Bilderbeek
}
