% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_beast2_pkg_names.R
\name{mrc_list}
\alias{mrc_list}
\title{Obsolete function to list all BEAST2 packages available and installed.
Use \link{get_beast2_pkg_names} instead}
\usage{
mrc_list()
}
\value{
a data frame
}
\description{
Obsolete function to list all BEAST2 packages available and installed.
Use \link{get_beast2_pkg_names} instead
}
\author{
Richèl J.C. Bilderbeek
}
