% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uninstall_beast2_pkg.R
\name{mrc_uninstall}
\alias{mrc_uninstall}
\title{Obsolete function to uninstall a BEAST2 package.
Use \link{uninstall_beast2_pkg} instead}
\usage{
mrc_uninstall(name = NA)
}
\arguments{
\item{name}{the package's name}
}
\value{
nothing. It does install the BEAST2 package
}
\description{
Obsolete function to uninstall a BEAST2 package.
Use \link{uninstall_beast2_pkg} instead
}
\author{
Richèl J.C. Bilderbeek
}
