% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uninstall_beast2_pkg.R
\name{uninstall_beast2_pkg}
\alias{uninstall_beast2_pkg}
\title{Uninstall a BEAST2 package}
\usage{
uninstall_beast2_pkg(name)
}
\arguments{
\item{name}{the package's name}
}
\value{
nothing. It does install the BEAST2 package
}
\description{
Uninstall a BEAST2 package
}
\author{
Richèl J.C. Bilderbeek
}
